move component update contexts to own file
