+++ packages/modules/alpha_ess/device.py
@@ -7,7 +7,7 @@ from typing import Dict, List, Union
 from helpermodules import log
 from modules.common import modbus
 from modules.common.abstract_device import AbstractDevice
-from modules.common.component_state import SingleComponentUpdateContext
+from modules.common.component_context import SingleComponentUpdateContext
 from modules.alpha_ess import bat
 from modules.alpha_ess import counter
 from modules.alpha_ess import inverter
